Remove deprecated tree mapping
