+++ docs/pages/mappings.rst
@@ -375,121 +375,6 @@ It permutes the array indices before passing the index information to the inner 
     view(1, 2, 3); // will pass {3, 1, 2} to inner mapping
 
 
-.. _label-tree-mapping:
-
-Tree (deprecated)
------------------
-
-WARNING: The tree mapping is currently not maintained and we consider deprecation.
-
-The LLAMA tree mapping is one approach to achieve the goal of mixing different mapping approaches.
-Furthermore, it tries to establish a general mapping description language and mapping definition framework.
-Let's take the example record dimension from the :ref:`dimensions section<label-dimensions>`:
-
-.. image:: ../images/layout_tree.svg
-
-As already mentioned this is a compile time tree. The idea of the tree mapping
-is now to extend this model to a compile time tree with run time annotations
-representing the repetition of branches and to define tree operations which
-create new trees out of the old ones while providing methods to translate tree
-coordinates from one tree to another.
-
-This is best demonstrated by an example.
-First of all the array dimensions needs to be represented as such an tree too.
-Let's assume array dimensions of :math:`128 \times 64`:
-
-.. image:: ../images/ud_tree_2.svg
-
-The record dimension is already a tree, but as it has no run time influence, only
-:math:`1` is annotated for these tree nodes:
-
-.. image:: ../images/layout_tree_2.svg
-
-Now the two trees are connected so that we can represent array and record
-dimensions with one tree:
-
-.. image:: ../images/start_tree_2.svg
-
-The mapping works now in this way that the tree is "flattened" from left to
-right using a breadth first traversal. Annotations represent repetitions of the node
-branches. So for this tree we would copy the record dimension :math:`64` times and
-:math:`128` times again -- basically this results in an array of struct
-approach, which is most probably not desired.
-
-So we want to transform the tree before flattening it. A struct of array
-approach may look like this:
-
-.. image:: ../images/soa_tree_2.svg
-
-Struct of array but with a padding after each 1024 elements may look like this:
-
-.. image:: ../images/padding_tree_2.svg
-
-The size of the leaf type in "pad" of course needs to be determined based on the
-desired alignment and sub tree sizes.
-
-Such a tree (with smaller array dimensions for easier drawing) …
-
-.. image:: ../images/example_tree.svg
-
-… may look like this mapped to memory:
-
-.. image:: ../images/example_mapping.svg
-
-In code a tree mapping is defined as :cpp:`llama::mapping::tree::Mapping`, but
-takes one more template parameter for the type of a tuple of tree operations and
-a further constructor parameter for the instantiation of this tuple.
-
-.. code-block:: C++
-
-    auto treeOperationList = llama::Tuple{
-        llama::mapping::tree::functor::LeafOnlyRT()
-    };
-
-    using Mapping = llama::mapping::tree::Mapping<
-        ArrayExtents,
-        RecordDim,
-        decltype(treeOperationList)
-    >;
-
-    Mapping mapping(
-        extents,
-        treeOperationList
-    );
-
-    // or using CTAD and an unused argument for the record dimension:
-    llama::mapping::tree::Mapping mapping(
-        extents,
-        llama::Tuple{
-            llama::mapping::tree::functor::LeafOnlyRT()
-        },
-        RecordDim{}
-    );
-
-The following tree operations are defined:
-
-Idem
-^^^^
-:cpp:`llama::mapping::tree::functor::Idem` does not change the tree at all.
-Basically a test functor for testing, how much the number of tree operations
-has an influence on the run time.
-
-LeafOnlyRT
-^^^^^^^^^^^
-:cpp:`llama::mapping::tree::functor::LeafOnlyRT` moves all run time parts of
-the tree to the leaves, basically creates a struct of array as seen above.
-However unlike :cpp:`llama::mapping::SoA` a combination with other mapping would
-be possible.
-
-MoveRTDown
-^^^^^^^^^^
-:cpp:`llama::mapping::tree::functor::MoveRTDown` moves a runtime multiplier from a node identified by a tree coordinate one level downward.
-This effectively divides the annotation at the node by a given factor and multiplies the direct child nodes by this factor.
-
-MoveRTDownFixed
-^^^^^^^^^^^^^^^
-Same as MoveRTDown but with a compile time factor.
-
 Dump visualizations
 -------------------
