+++ tests/core.cpp
@@ -17,7 +17,7 @@ TEST_CASE("prettyPrintType")
 {
     auto str = prettyPrintType<Particle>();
 #ifdef _WIN32
-    llama::mapping::tree::internal::replaceAll(str, "__int64", "long");
+    replaceAll(str, "__int64", "long");
 #endif
     const auto* const ref = R"(llama::Record<
     llama::Field<
